requirements.txt
python-dotenv==1.0.0
instagrapi==1.19.8
aiogram==2.25.1